+++ mrb_doc/models/image.rb
@@ -75,6 +75,16 @@ class Image
     nil
   end
 
+  # Generates a new image of width by height in the specified colour.
+  # @param width [Integer]
+  # @param height [Integer]
+  # @param colour [Colour]
+  # @return [Image]
+  def self.generate(width:, height:, colour: RAYWHITE)
+    # src/mruby_integration/models/image.cpp
+    Image.new
+  end
+
   # Used for alerting the user the image was not found at the specified path
   class NotFound < StandardError; end
 end