+++ test/fixtures/models.rb
@@ -1,1 +1,2 @@
 require 'test/fixtures/models/camera2d'
+require 'test/fixtures/models/image'